refractoring card style in movie card
