+++ src/styles/CardStyle.jsx
@@ -1,0 +1,16 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+
+const useCardStyle = makeStyles({
+    root: {
+      width: '300px',
+      margin: '10px 0'
+    },
+    media: {
+        paddingTop: '56.25%',
+        height: '450px'
+    }
+  });
+
+
+  export default useCardStyle;
